requirements.txt
python-telegram-bot>=20.0
phonenumbers
python-dotenv
